hack-bsf-backend/requirements.txt
Django>=4.2
djangorestframework==3.15.2
mysql-connector-python